style: move request validator to server
